Project Intern/requirements.txt
flask
google-genai
mysql-connector-python
pypdf
python-docx
scikit-learn
pandas
joblib
